Refactor code duplication in mergers
