+++ common/src/main/java/net/crystalixs/core/common/translation/TranslationConfigMergeService.java
@@ -1,51 +1,15 @@
 package net.crystalixs.core.common.translation;
 
+import net.crystalixs.core.common.sync.AbstractNodeMergeService;
 import net.crystalixs.core.common.logging.ChangeSet;
 import org.spongepowered.configurate.ConfigurationNode;
 
-import java.util.ArrayDeque;
 import java.util.Deque;
 
-public final class TranslationConfigMergeService {
+public final class TranslationConfigMergeService extends AbstractNodeMergeService {
 
-    public ChangeSet merge(ConfigurationNode defaults, ConfigurationNode user) {
-        ChangeSet changeSet = new ChangeSet();
-        merge(defaults, user, new ArrayDeque<>(), changeSet);
-        return changeSet;
-    }
-
-    private void merge(ConfigurationNode defaults, ConfigurationNode user, Deque<Object> path, ChangeSet changes) {
-        if (defaults.isMap()) {
-            if (!user.isMap() && !user.virtual()) {
-                replace(user, defaults);
-                changes.replaced(path);
-                return;
-            }
-            for (var entry : defaults.childrenMap().entrySet()) {
-                Object key = entry.getKey();
-                ConfigurationNode defaultChild = entry.getValue();
-                ConfigurationNode userChild = user.node(key);
-
-                path.addLast(key);
-                if (userChild.virtual()) {
-                    replace(userChild, defaultChild);
-                    changes.added(path);
-                } else {
-                    merge(defaultChild, userChild, path, changes);
-                }
-                path.removeLast();
-            }
-
-            user.childrenMap().keySet().stream()
-                    .filter(key -> !defaults.childrenMap().containsKey(key))
-                    .forEachOrdered(key -> {
-                        path.addLast(key);
-                        user.removeChild(key);
-                        changes.removed(path);
-                        path.removeLast();
-                    });
-            return;
-        }
+    @Override
+    protected void mergeNonMap(ConfigurationNode defaults, ConfigurationNode user, Deque<Object> path, ChangeSet changes) {
         if (defaults.isList()) {
             if (user.virtual()) {
                 replace(user, defaults);
@@ -62,6 +26,7 @@ public final class TranslationConfigMergeService {
             changes.replaced(path);
             return;
         }
+
         if (user.virtual()) {
             replace(user, defaults);
             changes.added(path);
@@ -71,9 +36,5 @@ public final class TranslationConfigMergeService {
             replace(user, defaults);
             changes.replaced(path);
         }
-    }
-
-    private void replace(ConfigurationNode target, ConfigurationNode source) {
-        target.from(source);
     }
 }
